integrating box2d and adding more editor options

+++ engine/components/AGridLines.cpp
@@ -9,19 +9,19 @@ void AGridLines::editorUpdate(float dt)
     ACamera *camera = AWindow::getScene()->getCamera();
     glm::vec2 cameraPos = camera->position;
     glm::vec2 projectionSize = AWindow::getScene()->getCamera()->projectionSize;
-    int firstX = ((int)(cameraPos.x / GRID_WIDTH) - 1) * GRID_WIDTH;
-    int firstY = ((int)(cameraPos.y / GRID_HEIGHT) - 1) * GRID_HEIGHT;
+    float firstX = ((int)(cameraPos.x / GRID_WIDTH) - 1) * GRID_WIDTH;
+    float firstY = ((int)(cameraPos.y / GRID_HEIGHT) - 1) * GRID_HEIGHT;
     int numVLines = (int)((projectionSize.x * camera->getZoom()) / GRID_WIDTH) + 2;
     int numHLines = (int)((projectionSize.y * camera->getZoom()) / GRID_HEIGHT) + 2;
 
-    int height = (int)(projectionSize.y * camera->getZoom()) + GRID_HEIGHT * 2;
-    int width = (int)(projectionSize.x * camera->getZoom()) + GRID_WIDTH * 2;
+    float height = (int)(projectionSize.y * camera->getZoom()) + GRID_HEIGHT * 2;
+    float width = (int)(projectionSize.x * camera->getZoom()) + GRID_WIDTH * 2;
     
     int maxLines = (numVLines > numHLines) ? numVLines : numHLines;
     for(int i = 0; i < maxLines; ++i)
     {
-        int x = firstX + (GRID_WIDTH * i);
-        int y = firstY + (GRID_HEIGHT * i);
+        float x = firstX + (GRID_WIDTH * i);
+        float y = firstY + (GRID_HEIGHT * i);
 
         if(i < numVLines)
         {
